neural_style_app/requirements.txt
Flask
Pillow
torch
torchvision
matplotlib